refactoring: GraphQL
upgrade: GraphQL 4.4.0
