+++ CoreSharp.GraphQL.AspNetCore/CoreSharp.GraphQL.AspNetCore.csproj
@@ -32,7 +32,7 @@
 
     <ItemGroup>
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-        <PackageReference Include="GraphQL.NewtonsoftJson" Version="3.1.0" />
+        <PackageReference Include="GraphQL.NewtonsoftJson" Version="4.4.0" />
         <ProjectReference Include="..\CoreSharp.GraphQL\CoreSharp.GraphQL.csproj" />
     </ItemGroup>
