+++ CoreSharp.GraphQL/CoreSharp.GraphQL.csproj
@@ -17,8 +17,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="GraphQL" Version="3.1.0" />
-        <PackageReference Include="GraphQL-Parser" Version="5.3.3" />
+        <PackageReference Include="GraphQL" Version="4.4.0" />
+        <PackageReference Include="GraphQL-Parser" Version="7.1.0" />
         <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
         <PackageReference Include="System.Linq.Dynamic.Core" Version="1.2.5" />
     </ItemGroup>
