--- CoreSharp.GraphQL/EntityGraphType.cs
@@ -1,146 +1,0 @@
-using GraphQL.Utilities;
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Linq.Expressions;
-using System.Reflection;
-using CoreSharp.Common.Attributes;
-using CoreSharp.DataAccess;
-using GraphQL;
-using GraphQL.Types;
-using DefaultValueAttribute = System.ComponentModel.DefaultValueAttribute;
-
-namespace CoreSharp.GraphQL
-{
-    /// <summary>
-    /// Allows you to automatically register the necessary fields for the specified type.
-    /// Supports <see cref="DescriptionAttribute"/>, <see cref="ObsoleteAttribute"/>, <see cref="System.ComponentModel.DefaultValueAttribute"/> and <see cref="RequiredAttribute"/>.
-    /// </summary>
-    /// <typeparam name="TSourceType"></typeparam>
-    public class EntityGraphType<TSourceType> : ObjectGraphType<TSourceType>
-        //where TSourceType : IEntity
-    {
-        // ReSharper disable once StaticMemberInGenericType
-        // private static readonly Type BaseEntityType = typeof(IEntity);
-
-        /// <summary>
-        /// Creates a GraphQL type by specifying fields to exclude from registration.
-        /// </summary>
-        /// <param name="excludedProperties"> Expressions for excluding fields, for example 'o => o.Age'. </param>
-        public EntityGraphType(params Expression<Func<TSourceType, object>>[] excludedProperties)
-        {
-            foreach (var propertyInfo in GetRegisteredProperties())
-            {
-                if (excludedProperties?.Any(p => GetPropertyName(p) == propertyInfo.Name) == true)
-                {
-                    continue;
-                }
-
-                Field(
-                    type: propertyInfo.PropertyType.GetGraphTypeFromType(IsNullableProperty(propertyInfo)),
-                    name: propertyInfo.Name,
-                    description: propertyInfo.Description(),
-                    deprecationReason: propertyInfo.ObsoleteMessage()
-                ).DefaultValue = (propertyInfo.GetCustomAttributes(typeof(DefaultValueAttribute), false).FirstOrDefault() as DefaultValueAttribute)?.Value;
-            }
-        }
-
-        private static bool IsNullableProperty(PropertyInfo propertyInfo)
-        {
-            if (Attribute.IsDefined(propertyInfo, typeof(RequiredAttribute)))
-            {
-                return false;
-            }
-
-            if (!propertyInfo.PropertyType.IsValueType)
-            {
-                return true;
-            }
-
-            return propertyInfo.PropertyType.IsGenericType && propertyInfo.PropertyType.GetGenericTypeDefinition() == typeof(Nullable<>);
-        }
-
-        private static string GetPropertyName(Expression<Func<TSourceType, object>> expression)
-        {
-            if (expression.Body is MemberExpression m1)
-            {
-                return m1.Member.Name;
-            }
-
-            if (expression.Body is UnaryExpression u && u.Operand is MemberExpression m2)
-            {
-                return m2.Member.Name;
-            }
-
-            throw new NotSupportedException($"Unsupported type of expression: {expression.GetType().Name}");
-        }
-
-        private IEnumerable<PropertyInfo> GetRegisteredProperties()
-        {
-            return typeof(TSourceType)
-                .GetProperties(BindingFlags.Public | BindingFlags.Instance)
-                .Where(p => IsEnabledForRegister(p.PropertyType, true));
-        }
-
-        private static bool IsEnabledForRegister(Type propertyType, bool firstCall)
-        {
-            if (propertyType.GetCustomAttribute<IgnoreAttribute>() != null)
-            {
-                return false;
-            }
-
-            if (propertyType == typeof(string))
-            {
-                return true;
-            }
-
-            if (propertyType.IsValueType) {
-                return true; // TODO: requires discussion: Nullable<T>, enums, any struct
-            }
-
-            if (GraphTypeTypeRegistry.Contains(propertyType))
-            {
-                return true;
-            }
-
-            /*if (!BaseEntityType.IsAssignableFrom(propertyType))
-            {
-                return false;
-            }*/
-
-            if (firstCall)
-            {
-                var realType = GetRealType(propertyType);
-                if (realType != propertyType)
-                {
-                    return IsEnabledForRegister(realType, false);
-                }
-            }
-
-            return false;
-        }
-
-        private static Type GetRealType(Type propertyType)
-        {
-            if (propertyType.IsGenericType && propertyType.GetGenericTypeDefinition() == typeof(Nullable<>))
-            {
-                return propertyType.GetGenericArguments()[0];
-            }
-
-            if (propertyType.IsArray)
-            {
-                return propertyType.GetElementType();
-            }
-
-            if (propertyType != typeof(string) && typeof(IEnumerable).IsAssignableFrom(propertyType))
-            {
-                return propertyType.GetEnumerableElementType();
-            }
-
-            return propertyType;
-        }
-    }
-}
