+++ CoreSharp.GraphQL/QueryResolver.cs
@@ -35,7 +35,7 @@ namespace CoreSharp.GraphQL
         public object Resolve(IResolveFieldContext context)
         {
             var queryHandler = _container.GetInstance(_queryHandlerType);
-            var variableValue = context.Arguments?.ContainsKey("query") == true ? JsonConvert.SerializeObject(context.Arguments["query"], _jsonSerializerSettings) : "{}";
+            var variableValue = context.Arguments?.ContainsKey("query") == true ? JsonConvert.SerializeObject(context.Arguments["query"].Value, _jsonSerializerSettings) : "{}";
             var query = JsonConvert.DeserializeObject(variableValue, _queryType, _jsonSerializerSettings);
 
             if (query != null)
